requirements.txt
requests
nltk
Scrapy
beautifulsoup4
newspaper3k
lxml_html_clean